Refactored common code for TranslateCoding UDFs.
